Extract Steam Workshop DB Builder into DatabaseBuilder
Move Steam workshop DB builder logic out of MainContent into a dedicated app.utils.db_builder.DatabaseBuilder QObject.
